+++ Engine/Code/Engine/Math/MathUtils.hpp
@@ -84,7 +84,7 @@ bool IsInRange(float value, float minInclusive, float maxExclusive);
 float Min(const float& x, const float& y);
 int Min(const int& x, const int& y);
 int Max(const int& x, const int& y);
-int Max(const float& x, const float& y);
+float Max(const float& x, const float& y);
 
 
 //---------------
